refactor(frontend): split keys page modules
